♻️  Overall Refactor
Since the code is very messy and has a lot of unused code, for this
year's edition the team decided to start with a clean nextjs repo and
add on to achieve what we have with clean, refactored code.

Basically until we achieve that, this branch will work as the "main"
branch for the frontend team. Once we achieve that we will merge this
branch with the main.